claude-3-agent/requirements.txt
livekit>=0.9.0
livekit-api
livekit-agents>=0.4.0
livekit-plugins-deepgram>=0.2.0
livekit-plugins-elevenlabs>=0.2.0
python-dotenv
anthropic